Fix Confetti overflow using a proper approach
