+++ src/styles/globals.css
@@ -53,7 +53,7 @@
   }
 
   body {
-    @apply h-screen overflow-hidden bg-purple-800 text-white accent-lemon-400 dark:bg-black  dark:bg-none dark:text-gray-200;
+    @apply h-screen bg-purple-800 text-white accent-lemon-400 dark:bg-black  dark:bg-none dark:text-gray-200;
   }
   input[type="number"]::-webkit-outer-spin-button,
   input[type="number"]::-webkit-inner-spin-button,
